packages/cli: split diff command implementation into multiple files
